+++ src/torchjd/aggregation/_pcgrad.py
@@ -3,17 +3,18 @@ from typing import cast
 import torch
 from torch import Tensor
 
-from torchjd._linalg import PSDMatrix
+from torchjd.linalg import PSDMatrix
 
 from ._aggregator_bases import GramianWeightedAggregator
-from ._utils.non_differentiable import raise_non_differentiable_error
-from ._weighting_bases import GramianWeighting
+from ._mixins import _NonDifferentiable
+from ._weighting_bases import _GramianWeighting
 
 
-class PCGradWeighting(GramianWeighting):
+# Non-differentiable: weights are modified in-place during the gradient projection loop.
+class PCGradWeighting(_NonDifferentiable, _GramianWeighting):
     """
-    :class:`~torchjd.aggregation.GramianWeighting` giving the weights of
-    :class:`~torchjd.aggregation.PCGrad`.
+    :class:`~torchjd.aggregation.Weighting` [:class:`~torchjd.linalg.PSDMatrix`]
+    giving the weights of :class:`~torchjd.aggregation.PCGrad`.
     """
 
     def forward(self, gramian: PSDMatrix, /) -> Tensor:
@@ -46,7 +47,7 @@ class PCGradWeighting(GramianWeighting):
         return weights.to(device)
 
 
-class PCGrad(GramianWeightedAggregator):
+class PCGrad(_NonDifferentiable, GramianWeightedAggregator):
     """
     :class:`~torchjd.aggregation.GramianWeightedAggregator` as defined in Algorithm 1 of
     `Gradient Surgery for Multi-Task Learning <https://arxiv.org/pdf/2001.06782.pdf>`_.
@@ -56,6 +57,3 @@ class PCGrad(GramianWeightedAggregator):
 
     def __init__(self) -> None:
         super().__init__(PCGradWeighting())
-
-        # This prevents running into a RuntimeError due to modifying stored tensors in place.
-        self.register_full_backward_pre_hook(raise_non_differentiable_error)