+++ src/torchjd/aggregation/_sum.py
@@ -2,12 +2,12 @@ import torch
 from torch import Tensor
 
 from ._aggregator_bases import WeightedAggregator
-from ._weighting_bases import MatrixWeighting
+from ._weighting_bases import _MatrixWeighting
 
 
-class SumWeighting(MatrixWeighting):
+class SumWeighting(_MatrixWeighting):
     r"""
-    :class:`~torchjd.aggregation.MatrixWeighting` that gives the weights
+    :class:`~torchjd.aggregation.Weighting` [:class:`~torchjd.linalg.Matrix`] that gives the weights
     :math:`\begin{bmatrix} 1 & \dots & 1 \end{bmatrix}^T \in \mathbb{R}^m`.
     """